Cleaned up experimental code
